+++ src/main/java/com/inventory/detail/repository/LaptopRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inventory.detail.model.Laptop;
 
-public interface LaptopRepository extends JpaRepository<Laptop, Integer> {
+public interface LaptopRepository extends JpaRepository<Laptop, Long> {
 
 }
